move invariants functions to modules
